requirements.txt
requests
SQLAlchemy==1.4.22
psycopg2
pytest==6.2.4
PyJWT==2.1.0